cost is not showing up in edit field
