+++ src/scripts/poiForm.js
@@ -105,12 +105,11 @@ const poiForm = {
   },
 
   handleAddNewPoi() {
-    console.log("save button clicked");
     let inputPoiName = document.querySelector("#poi__name").value
     let inputPoiDescription = document.querySelector("#poi__description").value
     let inputPoiCost = document.querySelector("#poi__cost").value
     let inputPoiPlace = document.querySelector("#poi__place").value
-    let inputPoiReview = `You haven't reviewed ${inputPoiName} yet. Press the Review/Edit Button to share your thoughts!`
+    let inputPoiReview = `You haven't reviewed ${inputPoiName} yet. Press the Edit button to share your thoughts!`
 
     let newInterest =
     {
@@ -120,7 +119,6 @@ const poiForm = {
       placeId: inputPoiPlace,
       review: inputPoiReview
     }
-    console.log(newInterest)
     
     data.postNewInterest(newInterest)
     .then(response => {
